Correction de Q28_4.java (double cast DANS le scalb).

+++ src/ch/epfl/javelo/Q28_4.java
@@ -10,10 +10,10 @@ public final class Q28_4 {
     }
 
     public static double asDouble(int q28_4) {
-        return Math.scalb(q28_4, -4);
+        return Math.scalb((double) q28_4, -4);
     }
 
     public static float asFloat(int q28_4) {
-        return Math.scalb(q28_4, -4);
+        return Math.scalb((float) q28_4, -4);
     }
 }
